+++ internal/install/recipes/bundle_test.go
@@ -20,8 +20,10 @@ func TestBundle_ShouldAddRecipes(t *testing.T) {
 	}
 
 	bundle.AddRecipe(newRecipe)
+	index, found := bundle.ContainsName(newRecipe.Recipe.Name)
 	require.Equal(t, len(bundle.BundleRecipes), 2)
-	require.True(t, true, bundle.ContainsName(newRecipe.Recipe.Name))
+	require.Equal(t, true, found)
+	require.Equal(t, 1, index)
 }
 
 func TestBundle_ShouldNotUpdateRecipe(t *testing.T) {
@@ -37,7 +39,9 @@ func TestBundle_ShouldNotUpdateRecipe(t *testing.T) {
 	bundle.AddRecipe(bundleRecipe)
 
 	require.Equal(t, len(bundle.BundleRecipes), 1)
-	require.Equal(t, true, bundle.ContainsName(bundleRecipe.Recipe.Name))
+	_, found := bundle.ContainsName(bundleRecipe.Recipe.Name)
+
+	require.Equal(t, true, found)
 }
 
 func TestBundle_ShouldContainRecipeName(t *testing.T) {
@@ -48,8 +52,9 @@ func TestBundle_ShouldContainRecipeName(t *testing.T) {
 			},
 		},
 	}
+	_, found := bundle.ContainsName("recipe1")
 
-	require.Equal(t, true, bundle.ContainsName("recipe1"))
+	require.Equal(t, true, found)
 }
 
 func TestBundle_ShouldNotContainRecipeName(t *testing.T) {
@@ -60,8 +65,9 @@ func TestBundle_ShouldNotContainRecipeName(t *testing.T) {
 			},
 		},
 	}
+	_, found := bundle.ContainsName("some other name")
 
-	require.Equal(t, false, bundle.ContainsName("some other name"))
+	require.Equal(t, false, found)
 }
 
 func TestBundle_ShouldBeGuided(t *testing.T) {